Fix bug where search bar stops working after clicking NUbots logo (#221)
* Ensure search bar has rendered before running algolia initialiser

* Formatting

* Remove format check from static directory

* Move event listener back to static folder
